Update mouse mode/tool for spectral extraction
