+++ glue/viewers/common/qt/mouse_mode.py
@@ -581,60 +581,3 @@ class ContrastMode(MouseMode):
                 r.triggered.connect(nonpartial(self._move_callback, self))
 
         return result
-
-
-class SpectrumExtractorMode(RoiMode):
-
-    """
-    Let's the user select a region in an image and,
-    when connected to a SpectrumExtractorTool, uses this
-    to display spectra extracted from that position
-    """
-    persistent = True
-
-    icon = 'glue_spectrum'
-    mode_id = 'Spectrum'
-    action_text = 'Spectrum'
-    tool_tip = 'Extract a spectrum from the selection'
-    shortcut = 'S'
-
-
-    def __init__(self, viewer, **kwargs):
-        super(SpectrumExtractorMode, self).__init__(viewer, **kwargs)
-        self._roi_tool = qt_roi.QtRectangularROI(self._axes) # default
-
-    def menu_actions(self):
-        result = []
-
-        a = QtWidgets.QAction('Rectangle', None)
-        a.triggered.connect(nonpartial(self.set_roi_tool, 'Rectangle'))
-        result.append(a)
-
-        a = QtWidgets.QAction('Circle', None)
-        a.triggered.connect(nonpartial(self.set_roi_tool, 'Circle'))
-        result.append(a)
-
-        a = QtWidgets.QAction('Polygon', None)
-        a.triggered.connect(nonpartial(self.set_roi_tool, 'Polygon'))
-        result.append(a)
-
-        for r in result:
-            if self._move_callback is not None:
-                r.triggered.connect(nonpartial(self._move_callback, self))
-
-        return result
-
-    def set_roi_tool(self, mode):
-        if mode is 'Rectangle':
-            self._roi_tool = qt_roi.QtRectangularROI(self._axes)
-
-        if mode is 'Circle':
-            self._roi_tool = qt_roi.QtCircularROI(self._axes)
-
-        if mode is 'Polygon':
-            self._roi_tool = qt_roi.QtPolygonalROI(self._axes)
-
-        self._roi_tool.plot_opts.update(edgecolor='#c51b7d',
-                                        facecolor=None,
-                                        edgewidth=3,
-                                        alpha=1.0)
